Enable empty tile directories to be extracted more easily
'discernFromClassesToExtract' added as an option to otherClassNames argument for extractAnnotationTiles() function. This will cause all classes not found in annotations to be automatically created as empty tile and/or mask directories. Useful for WSI datasets where not all classes are present in every WSI, and the combination of classes present in each slide varies.